requirements.txt
ultralytics
numpy
pyyaml
opencv-contrib-python